imagenstable/requirements.txt
crewai
google-generativeai
stability-sdk
python-dotenv
pillow
requests
python-multipart
crewai-tools
ollama
langchain-community